+++ apps/web/package.json
@@ -51,6 +51,7 @@
     "embla-carousel-react": "^8.6.0",
     "i18next": "^25.5.3",
     "input-otp": "^1.4.2",
+    "js-cookie": "^3.0.5",
     "lucide-react": "^0.544.0",
     "next-themes": "^0.4.6",
     "nuqs": "^2.7.0",
@@ -61,6 +62,7 @@
     "react-i18next": "^16.0.0",
     "react-is": "^19.2.0",
     "react-resizable-panels": "^3.0.6",
+    "react-use": "^17.6.0",
     "recharts": "^2.15.4",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1",
@@ -74,6 +76,7 @@
     "@tailwindcss/vite": "^4.1.14",
     "@tanstack/react-router-devtools": "^1.132.33",
     "@tanstack/router-plugin": "^1.132.33",
+    "@types/js-cookie": "^3.0.6",
     "@types/node": "^24.6.2",
     "@types/react": "^19.2.0",
     "@types/react-dom": "^19.2.0",
