+++ apps/web/src/components/pages/SlaveNodes/index.tsx
@@ -1,0 +1,4 @@
+import SystemConfig from './SystemConfig';
+import SlaveNodes from './SlaveNodes';
+
+export { SystemConfig, SlaveNodes };
