+++ apps/web/src/services/slave.service.ts
@@ -1,7 +1,5 @@
-import axios from 'axios';
+import api from './api';
 import { SlaveNode } from '@/types';
-
-const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:3001/api';
 
 export interface RegisterSlaveNodeRequest {
   name: string;
@@ -39,39 +37,22 @@ export interface SlaveNodeWithLogs extends SlaveNode {
   syncLogs?: SyncLog[];
 }
 
-// Helper function to get headers
-const getHeaders = () => {
-  const token = localStorage.getItem('accessToken');
-  return {
-    'Content-Type': 'application/json',
-    Authorization: token ? `Bearer ${token}` : '',
-  };
-};
-
 class SlaveNodeService {
   async getAll(): Promise<SlaveNode[]> {
-    const response = await axios.get(`${API_URL}/slave/nodes`, {
-      headers: getHeaders(),
-    });
+    const response = await api.get('/slave/nodes');
     return response.data.data;
   }
 
   async getById(id: string): Promise<SlaveNodeWithLogs> {
-    const response = await axios.get(`${API_URL}/slave/nodes/${id}`, {
-      headers: getHeaders(),
-    });
+    const response = await api.get(`/slave/nodes/${id}`);
     return response.data.data;
   }
 
   async register(data: RegisterSlaveNodeRequest) {
     console.log('SlaveNodeService.register called with:', data);
-    console.log('API_URL:', API_URL);
-    console.log('Headers:', getHeaders());
-    
+
     try {
-      const response = await axios.post(`${API_URL}/slave/nodes`, data, {
-        headers: getHeaders(),
-      });
+      const response = await api.post('/slave/nodes', data);
       console.log('Register response:', response.data);
       return response.data;
     } catch (error: any) {
@@ -81,52 +62,39 @@ class SlaveNodeService {
   }
 
   async update(id: string, data: UpdateSlaveNodeRequest) {
-    const response = await axios.put(`${API_URL}/slave/nodes/${id}`, data, {
-      headers: getHeaders(),
-    });
+    const response = await api.put(`/slave/nodes/${id}`, data);
     return response.data;
   }
 
   async delete(id: string) {
-    const response = await axios.delete(`${API_URL}/slave/nodes/${id}`, {
-      headers: getHeaders(),
-    });
+    const response = await api.delete(`/slave/nodes/${id}`);
     return response.data;
   }
 
   async syncToNode(id: string, data: SyncConfigRequest = {}) {
-    const response = await axios.post(`${API_URL}/slave/nodes/${id}/sync`, data, {
-      headers: getHeaders(),
-    });
+    const response = await api.post(`/slave/nodes/${id}/sync`, data);
     return response.data;
   }
 
   async syncToAll() {
-    const response = await axios.post(`${API_URL}/slave/nodes/sync-all`, {}, {
-      headers: getHeaders(),
-    });
+    const response = await api.post('/slave/nodes/sync-all', {});
     return response.data;
   }
 
   async getStatus(id: string) {
-    const response = await axios.get(`${API_URL}/slave/nodes/${id}/status`, {
-      headers: getHeaders(),
-    });
+    const response = await api.get(`/slave/nodes/${id}/status`);
     return response.data;
   }
 
   async getSyncHistory(id: string, limit: number = 50) {
-    const response = await axios.get(`${API_URL}/slave/nodes/${id}/sync-history`, {
-      headers: getHeaders(),
+    const response = await api.get(`/slave/nodes/${id}/sync-history`, {
       params: { limit },
     });
     return response.data.data;
   }
 
   async regenerateApiKey(id: string) {
-    const response = await axios.post(`${API_URL}/slave/nodes/${id}/regenerate-key`, {}, {
-      headers: getHeaders(),
-    });
+    const response = await api.post(`/slave/nodes/${id}/regenerate-key`, {});
     return response.data;
   }
 }
