+++ apps/web/src/services/user.service.ts
@@ -1,14 +1,4 @@
-import axios from 'axios';
-
-const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:3001/api';
-
-const getAuthHeaders = () => {
-  const token = localStorage.getItem('accessToken');
-  return {
-    'Content-Type': 'application/json',
-    'Authorization': token ? `Bearer ${token}` : ''
-  };
-};
+import api from './api';
 
 export interface User {
   id: string;
@@ -66,66 +56,49 @@ export interface UserStats {
 const userService = {
   // Get all users
   async getAll(params?: { role?: string; status?: string; search?: string }): Promise<{ success: boolean; data: User[] }> {
-    const queryString = params ? new URLSearchParams(params as any).toString() : '';
-    const response = await axios.get(`${API_URL}/users${queryString ? `?${queryString}` : ''}`, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.get('/users', { params });
     return response.data;
   },
 
   // Get single user
   async getById(id: string): Promise<{ success: boolean; data: User }> {
-    const response = await axios.get(`${API_URL}/users/${id}`, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.get(`/users/${id}`);
     return response.data;
   },
 
   // Create new user
   async create(data: CreateUserData): Promise<{ success: boolean; data: User; message: string }> {
-    const response = await axios.post(`${API_URL}/users`, data, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.post('/users', data);
     return response.data;
   },
 
   // Update user
   async update(id: string, data: UpdateUserData): Promise<{ success: boolean; data: User; message: string }> {
-    const response = await axios.put(`${API_URL}/users/${id}`, data, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.put(`/users/${id}`, data);
     return response.data;
   },
 
   // Delete user
   async delete(id: string): Promise<{ success: boolean; message: string }> {
-    const response = await axios.delete(`${API_URL}/users/${id}`, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.delete(`/users/${id}`);
     return response.data;
   },
 
   // Toggle user status
   async updateStatus(id: string, status: 'active' | 'inactive' | 'suspended'): Promise<{ success: boolean; data: User; message: string }> {
-    const response = await axios.patch(`${API_URL}/users/${id}/status`, { status }, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.patch(`/users/${id}/status`, { status });
     return response.data;
   },
 
   // Reset user password
   async resetPassword(id: string): Promise<{ success: boolean; message: string; data?: any }> {
-    const response = await axios.post(`${API_URL}/users/${id}/reset-password`, {}, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.post(`/users/${id}/reset-password`, {});
     return response.data;
   },
 
   // Get user statistics
   async getStats(): Promise<{ success: boolean; data: UserStats }> {
-    const response = await axios.get(`${API_URL}/users/stats`, {
-      headers: getAuthHeaders()
-    });
+    const response = await api.get('/users/stats');
     return response.data;
   }
 };
